Export simplifyAPIDataForSingleState to utilities
