+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/index.tsx":"1","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/App/App.tsx":"2","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/apiCalls.tsx":"3","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/USAState/USAState.tsx":"4","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/Stat/Stat.tsx":"5","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/comparisonData.tsx":"6","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonContainer/ComparisonContainer.tsx":"7","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonDetails/ComparisonDetails.tsx":"8","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonCard/ComparisonCard.tsx":"9"},{"size":280,"mtime":1610222194368,"results":"10","hashOfConfig":"11"},{"size":7687,"mtime":1610405642900,"results":"12","hashOfConfig":"11"},{"size":601,"mtime":1610048850209,"results":"13","hashOfConfig":"11"},{"size":354,"mtime":1610222194367,"results":"14","hashOfConfig":"11"},{"size":470,"mtime":1610222194366,"results":"15","hashOfConfig":"11"},{"size":1247,"mtime":1610322672556,"results":"16","hashOfConfig":"11"},{"size":2240,"mtime":1610382820430,"results":"17","hashOfConfig":"11"},{"size":2595,"mtime":1610405777395,"results":"18","hashOfConfig":"11"},{"size":159,"mtime":1610405377337,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"17bwe72",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"29","usedDeprecatedRules":"22"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"40","usedDeprecatedRules":"22"},"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/index.tsx",[],["41","42"],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/App/App.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/apiCalls.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/USAState/USAState.tsx",["43"],"import './USAState.scss';\n\nconst USAState = (prop: { date?: string }) => {\n  return(\n    <section>\n      <section className='title-info'>\n        <h1 className='state-image'></h1>\n        <h2 className='state-name'>Colorado</h2>\n      </section>\n      <h3 className='date'>Information for { prop.date }</h3>\n    </section>\n  )\n}\n\nexport default USAState;","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/Stat/Stat.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/comparisonData.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonContainer/ComparisonContainer.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonDetails/ComparisonDetails.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/ComparisonCard/ComparisonCard.tsx",["44"],"import './ComparisonCard.scss';\n\n\nconst ComparisonCard = (prop: { }) => {\n  const generateComparisonCards = () => {\n    \n  }\n}\n\n\nexport default ComparisonCard;",{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},{"ruleId":"49","severity":1,"message":"50","line":7,"column":9,"nodeType":"51","endLine":7,"endColumn":37},{"ruleId":"52","severity":1,"message":"53","line":5,"column":9,"nodeType":"54","messageId":"55","endLine":5,"endColumn":32},"no-native-reassign",["56"],"no-negated-in-lhs",["57"],"jsx-a11y/heading-has-content","Headings must have content and the content must be accessible by a screen reader.","JSXOpeningElement","@typescript-eslint/no-unused-vars","'generateComparisonCards' is assigned a value but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
+[{"/Users/karacaputo/turing/3module/covid_comparisons/src/index.tsx":"1","/Users/karacaputo/turing/3module/covid_comparisons/src/App/App.tsx":"2","/Users/karacaputo/turing/3module/covid_comparisons/src/apiCalls.tsx":"3","/Users/karacaputo/turing/3module/covid_comparisons/src/Stat/Stat.tsx":"4","/Users/karacaputo/turing/3module/covid_comparisons/src/comparisonData.tsx":"5","/Users/karacaputo/turing/3module/covid_comparisons/src/USAState/USAState.tsx":"6","/Users/karacaputo/turing/3module/covid_comparisons/src/ComparisonContainer/ComparisonContainer.tsx":"7","/Users/karacaputo/turing/3module/covid_comparisons/src/ComparisonDetails/ComparisonDetails.tsx":"8"},{"size":280,"mtime":1610232598072,"results":"9","hashOfConfig":"10"},{"size":8399,"mtime":1610412171060,"results":"11","hashOfConfig":"10"},{"size":601,"mtime":1610053584082,"results":"12","hashOfConfig":"10"},{"size":470,"mtime":1610068395852,"results":"13","hashOfConfig":"10"},{"size":2057,"mtime":1610410462295,"results":"14","hashOfConfig":"10"},{"size":354,"mtime":1610068395853,"results":"15","hashOfConfig":"10"},{"size":2240,"mtime":1610402230142,"results":"16","hashOfConfig":"10"},{"size":2595,"mtime":1610410462295,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"118sn6p",{"filePath":"20","messages":"21","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/karacaputo/turing/3module/covid_comparisons/src/index.tsx",[],"/Users/karacaputo/turing/3module/covid_comparisons/src/App/App.tsx",["34"],"/Users/karacaputo/turing/3module/covid_comparisons/src/apiCalls.tsx",[],"/Users/karacaputo/turing/3module/covid_comparisons/src/Stat/Stat.tsx",[],"/Users/karacaputo/turing/3module/covid_comparisons/src/comparisonData.tsx",[],"/Users/karacaputo/turing/3module/covid_comparisons/src/USAState/USAState.tsx",["35"],"/Users/karacaputo/turing/3module/covid_comparisons/src/ComparisonContainer/ComparisonContainer.tsx",[],"/Users/karacaputo/turing/3module/covid_comparisons/src/ComparisonDetails/ComparisonDetails.tsx",[],{"ruleId":null,"fatal":true,"severity":2,"message":"36","line":11,"column":0},{"ruleId":"37","severity":1,"message":"38","line":7,"column":9,"nodeType":"39","endLine":7,"endColumn":37},"Parsing error: 'from' expected.","jsx-a11y/heading-has-content","Headings must have content and the content must be accessible by a screen reader.","JSXOpeningElement"]
